Adds additional workbox configuration.
It appears that Workbox v3-v4 suffer from an issue regarding the detection
of hashes contained within webpack-generated assets. Though the option of
providing a RegExp pattern with which to excluding revisioning is provided,
it apparently only applies to non-webpack pipeline assets.

There's not much information available regarding this issue and the author
makes no claim to have a sophisticated understanding of PWA or workbox.